+++ hart/exceptions.py
@@ -1,0 +1,2 @@
+class UserError(Exception):
+    '''Raised when a user gives invalid instructions.'''