+++ tests/conftest.py
@@ -4,10 +4,15 @@ import tempfile
 import pytest
 
 
-@pytest.yield_fixture
+@pytest.fixture
 def named_tempfile():
     tmp = tempfile.NamedTemporaryFile(delete=False)
     try:
         yield tmp
     finally:
         os.remove(tmp.name)
+
+
+@pytest.fixture
+def local_config():
+    return os.path.join(os.path.dirname(__file__), '..', 'hart.toml')